Devices: clean warning after replace of int with size_t in Image.

+++ src/devices/realsense2/realsense2Driver.cpp
@@ -1495,13 +1495,13 @@ bool realsense2Driver::getImage(yarp::sig::ImageOf<yarp::sig::PixelMono>& image)
         return false;
     }
 
-    int singleImage_rowSizeByte   =  image.getRowSize()/2;
+    size_t singleImage_rowSizeByte   =  image.getRowSize()/2;
     unsigned char * pixelLeft     =  (unsigned char*) (frm1.get_data());
     unsigned char * pixelRight    =  (unsigned char*) (frm2.get_data());
     unsigned char * pixelOutLeft  =  image.getRawImage();
     unsigned char * pixelOutRight =  image.getRawImage() + singleImage_rowSizeByte;
 
-    for (int h=0; h< image.height(); h++)
+    for (size_t h=0; h< image.height(); h++)
     {
         memcpy(pixelOutLeft, pixelLeft, singleImage_rowSizeByte);
         memcpy(pixelOutRight, pixelRight, singleImage_rowSizeByte);
